be/int/web/receptions: merges files for custom response writer implementation
