+++ backend/internal/web/reception/receptionist.go
@@ -18,6 +18,16 @@ import (
 	"time"
 )
 
+type response struct {
+	http.ResponseWriter
+	Status int
+}
+
+func (r *response) WriteHeader(statusCode int) {
+	r.Status = statusCode
+	r.ResponseWriter.WriteHeader(statusCode)
+}
+
 type RequestId string
 
 type receptionist struct {
