--- backend/internal/web/reception/response.go
@@ -1,13 +1,0 @@
-package reception
-
-import "net/http"
-
-type response struct {
-	http.ResponseWriter
-	Status int
-}
-
-func (r *response) WriteHeader(statusCode int) {
-	r.Status = statusCode
-	r.ResponseWriter.WriteHeader(statusCode)
-}
